requirements.txt
cairocffi==0.8.0
certifi==2017.7.27.1
cffi==1.11.2
chardet==3.0.4
cycler==0.10.0
idna==2.6
matplotlib==2.1.0
numpy==1.13.3
oauthlib==2.0.6
pycparser==2.18
pyparsing==2.2.0
python-dateutil==2.6.1
python-igraph==0.7.1.post6
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
scipy==1.0.0
six==1.11.0
tweepy==3.5.0
urllib3==1.22
